blockmanager: Remove dependency on cfg in main package.
